consolidate Digest/RecoverSigner into service/attestor/evm alongside SignABI, removing the duplicated domain-separation tags and digest math (dmitry)
