+++ link/internal/relay/proofgen/attestation/quorum_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/cosmos/ibc/link/internal/service/attestor"
+	attestorevm "github.com/cosmos/ibc/link/internal/service/attestor/evm"
 	"github.com/cosmos/ibc/link/internal/tests/mocks"
 	v2 "github.com/cosmos/ibc/link/internal/types/v2"
 )
@@ -24,7 +25,7 @@ func signedAttestor(t *testing.T, name string, attestedData []byte) *attestor.Mo
 	key, err := crypto.GenerateKey()
 	require.NoError(t, err)
 
-	digest := Digest(TagStateAttestation, attestedData)
+	digest := attestorevm.Digest(attestorevm.TagStateAttestation, attestedData)
 	sig, err := crypto.Sign(digest[:], key)
 	require.NoError(t, err)
 
@@ -128,7 +129,7 @@ func TestQueryQuorum(t *testing.T) {
 		key, err := crypto.GenerateKey()
 		require.NoError(t, err)
 
-		digest := Digest(TagStateAttestation, data)
+		digest := attestorevm.Digest(attestorevm.TagStateAttestation, data)
 		sig, err := crypto.Sign(digest[:], key)
 		require.NoError(t, err)
