+++ link/internal/service/attestor/evm/signing.go
@@ -4,6 +4,9 @@ import (
 	"context"
 	"crypto/sha256"
 	"fmt"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
 )
 
 // Signer signs a 32-byte ECDSA digest.
@@ -17,14 +20,21 @@ const (
 	TagPacketAttestation byte = 0x02
 )
 
-// SignABI signs sha256(tag || sha256(data)), matching the domain separation.
-func SignABI(ctx context.Context, signer Signer, tag byte, data []byte) ([]byte, error) {
-	innerHash := sha256.Sum256(data)
+// Digest computes sha256(tag || sha256(data)), the domain-separated digest
+// AttestationLightClient.sol verifies signatures over.
+func Digest(tag byte, data []byte) [32]byte {
+	inner := sha256.Sum256(data)
 
 	var signingInput [1 + sha256.Size]byte
 	signingInput[0] = tag
-	copy(signingInput[1:], innerHash[:])
-	digest := sha256.Sum256(signingInput[:])
+	copy(signingInput[1:], inner[:])
+
+	return sha256.Sum256(signingInput[:])
+}
+
+// SignABI signs Digest(tag, data).
+func SignABI(ctx context.Context, signer Signer, tag byte, data []byte) ([]byte, error) {
+	digest := Digest(tag, data)
 
 	signature, err := signer.Sign(ctx, digest[:])
 	if err != nil {
@@ -51,4 +61,30 @@ func normalizeSignature(signature []byte) ([]byte, error) {
 	}
 
 	return normalized, nil
+}
+
+// RecoverSigner recovers the address that produced sig (65-byte r||s||v)
+// over digest, the inverse of SignABI: it accepts either recovery-ID form
+// (crypto.SigToPub requires the signer's raw 0/1 form, but attestors sign
+// with Ethereum's legacy v of 27/28 via normalizeSignature above, and a
+// verifier only ever sees the signature on the wire, never picks the form
+// itself).
+func RecoverSigner(digest [32]byte, sig []byte) (common.Address, error) {
+	if len(sig) != 65 {
+		return common.Address{}, fmt.Errorf("signature must be 65 bytes, got %d", len(sig))
+	}
+
+	normalized := make([]byte, 65)
+	copy(normalized, sig)
+
+	if normalized[64] >= 27 {
+		normalized[64] -= 27
+	}
+
+	pub, err := crypto.SigToPub(digest[:], normalized)
+	if err != nil {
+		return common.Address{}, fmt.Errorf("recovering signer public key: %w", err)
+	}
+
+	return crypto.PubkeyToAddress(*pub), nil
 }
